experiment/2-DeepSeek使用与Prompt工程/pyproject.toml
[project]
name = "day02"
version = "0.1.0"
description = "DeepSeek使用与Prompt工程"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "dashscope>=1.24.2",
    "fastapi>=0.116.1",
    "modelscope>=1.29.1",
    "peft>=0.17.1",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "transformers>=4.56.0",
    "uvicorn>=0.35.0",
]
